pyproject.toml
[project]
name = "rag-qa-system"
version = "0.1.0"
description = "RAG Q&A System - FastAPI application for document Q&A using Retrieval-Augmented Generation"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "vinaygandhi", email = "vgandhi1@binghamton.edu" }
]
keywords = ["rag", "qa", "fastapi", "langchain", "qdrant", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    # Core Framework
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",

    # LangChain & AI
    "langchain",
    "langchain-openai",
    "langchain-qdrant",
    "langchain-community",
    "langchain-text-splitters",

    # Vector Database
    "qdrant-client",

    # Document Processing
    "pypdf",
    "python-docx",

    # Configuration & Validation
    "pydantic",
    "pydantic-settings",
    "python-dotenv",

    # Logging & Monitoring
    "structlog",

    # HTTP Client
    "httpx",

    # Evaluation
    "ragas>=0.1.0",
    "datasets>=2.14.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",

    # Linting & Formatting
    "ruff>=0.2.0",
    "black>=24.1.0",

    # Type Checking
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/rag-qa-project"
Documentation = "https://github.com/yourusername/rag-qa-project#readme"
Repository = "https://github.com/yourusername/rag-qa-project"
Issues = "https://github.com/yourusername/rag-qa-project/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=60",
    "--asyncio-mode=auto",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # raise from err (stylistic choice for exception handling)
    "UP028", # yield from (stylistic choice)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["ARG", "F841"]  # Allow unused arguments and variables in tests
"app/main.py" = ["ARG001"]   # Allow unused lifespan app argument (required by FastAPI)
"app/core/rag_chain.py" = ["ARG002"]  # Allow unused arguments for API consistency

[tool.ruff.lint.isort]
known-first-party = ["app"]

# Black Configuration
[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# MyPy Configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "qdrant_client.*",
    "pypdf.*",
    "docx.*",
]
ignore_missing_imports = true
